+++ example/android/app/src/main/kotlin/com/fl/pip/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.fl.pip.example
-
-import com.fl.pip.FlPiPActivity
-
-class MainActivity : FlPiPActivity() {}