+++ example/android/app/src/main/kotlin/fl/pip/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package fl.pip.example
+
+import fl.pip.FlPiPActivity
+
+class MainActivity : FlPiPActivity() {}